Repeated habit couldn't be be written again
Problem:
Addhabit() is comparing all the systems from all days with the input value.
It means that if you delete a habit and then want to add it again, it's
not possible.

Solution:
AddHabit() Compare input value with today systems.
This allows you to be able to add the habit again and keep checking that
you are not repeating systems in a day.